refactor: use job_id, rename files to remove clutter
