+++ frappe_azure_blob_storage/commands/__init__.py
@@ -1,3 +1,3 @@
-from .azure_commands import migrate_azure_files
+from .azure import migrate_azure_files
 
 commands = [migrate_azure_files]
