+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "frappe_azure_blob_storage"
 authors = [
-    { name = "rtCamp", email = "erp@rtcamp.com"}
+    { name = "rtCamp", email = "frappe@rtcamp.com"}
 ]
 description = "Blob Storage support for Azure"
 requires-python = ">=3.10"
@@ -23,8 +23,11 @@ build-backend = "flit_core.buildapi"
 # package_name = "~=1.1.0"
 
 [tool.ruff]
-line-length = 110
+line-length = 120
 target-version = "py310"
+exclude = [
+    "**/doctype/*/boilerplate/*.py" # boilerplate are template strings, not valid python
+]
 
 [tool.ruff.lint]
 select = [
@@ -41,15 +44,16 @@ ignore = [
     "B018", # useless expression, not assigned to anything
     "B023", # function doesn't bind loop variable - will have last iteration's value
     "B904", # raise inside except without from
-    "E101", # indentation contains mixed spaces and tabs
     "E402", # module level import not at top of file
     "E501", # line too long
     "E741", # ambiguous variable name
-    "F401", # "unused" imports
     "F403", # can't detect undefined names from * import
     "F405", # can't detect undefined names from * import
     "F722", # syntax error in forward type annotation
     "W191", # indentation contains tabs
+    "RUF001", # string contains ambiguous unicode character
+    "UP032", # Use f-string instead of `format` call (translations)
+    "UP030", # Use implicit references for positional format fields (translations)
 ]
 typing-modules = ["frappe.types.DF"]
